Moving geocoder to api calls
